+++ CHANGELOG.md
@@ -14,6 +14,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+-   Added test for invalid agent sessionid (@Cx01N)
+
+### Fixed
+-   Fixed issue that invalid session IDs were accepted by the server (@Cx01N)
+
 ## [5.9.2] - 2024-01-31
 
 ### Fixed